requirements.txt
opencv-python==4.8.1.78
numpy==1.24.3
streamlit==1.28.1
Pillow==10.0.1 